add (u64) explicit cast to the argument of cpu_to_le64(sector)
- dm-crypt does like this.

+++ src/dm-writeboost-metadata.c
@@ -1208,7 +1208,7 @@ void prepare_segment_header_device(void *rambuffer,
 		struct metablock *mb = src->mb_array + i;
 		struct metablock_device *mbdev = dest->mbarr + i;
 
-		mbdev->sector = cpu_to_le64(mb->sector);
+		mbdev->sector = cpu_to_le64((u64)mb->sector);
 		mbdev->dirty_bits = mb->dirty_bits;
 	}
